+++ src/gui/painting/qpainter.cpp
@@ -197,9 +197,9 @@ void QPainterPrivate::checkEmulation()
 
     if (state->brush.style() == Qt::TexturePattern) {
         if (qHasPixmapTexture(state->brush))
-            doEmulation |= !qFuzzyCompare(state->brush.texture().devicePixelRatioF(), 1.0);
+            doEmulation |= !qFuzzyCompare(state->brush.texture().devicePixelRatioF(), qreal(1.0));
         else
-            doEmulation |= !qFuzzyCompare(state->brush.textureImage().devicePixelRatioF(), 1.0);
+            doEmulation |= !qFuzzyCompare(state->brush.textureImage().devicePixelRatioF(), qreal(1.0));
     }
 
     if (doEmulation && extended->flags() & QPaintEngineEx::DoNotEmulate)
